refac: refactor of postgrestest package
